- Use RateLimiter for Smarti requests - Use propagation function for 'onMessage' and 'onClose' - Make Smarti the default knowledge provider - Reordered Smarti backend settings - Added descriptions for Smarti backend settings - Also reload the settings, when reloading the Smarti widget - Refactored file names
